requirements.txt
# Core dependencies
fastapi==0.128.7
uvicorn==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# SQLAlchemy and database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0  # For SQLite (development)
asyncpg==0.30.0    # For PostgreSQL (production)

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.0
